Add Firefox support (geckodriver) (#2)
## Summary

This adds geckodriver as an available driver.

The hardest part was definitely trying to figure out how to get the tar file extracted. My problem kept ended up being that Github was returning a redirect and I didn't know. I'm not loving the organization of the install files for the drivers.

## Added Dependencies

### [Halite](https://github.com/icyleaf/halite)

I started using the Halite http client because I have used it before and it has automatic redirect handling (hunting down why it worked and my other code didn't was how I figured out the redirect issue). This is a dependency that could easily be removed in the future now that I know about the redirects.

### [Crystar](https://github.com/naqvis/crystar)

Adding this dependency to extract the tar file. I was also considering [tarball.cr](https://github.com/arcage/tarball.cr) which has built in gzip support, but I went with Crystar because it had more stars, more recent commits, and seemed to be more in line with the gzip api.